+++ app/src/main/res/layout/activity_login_success.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.cookandroid.teamproject1.id.view.LoginSuccessActivity"
+    tools:context="com.cookandroid.teamproject1.id.view.activity.LoginSuccessActivity"
     android:background="#6BBFEE"
     >
