+++ app/src/main/res/layout/create_account.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FFFCF6"
-    tools:context=".id.view.CreateAccountActivity">
+    tools:context=".id.view.activity.CreateAccountActivity">
 
     <ImageView
         android:id="@+id/create_account_id_checkimage"
