+++ app/src/main/res/layout/intro_activity.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.cookandroid.teamproject1.id.view.IntroActivity"
+    tools:context="com.cookandroid.teamproject1.id.view.activity.IntroActivity"
     android:background="#6BBFEE"
     >
