+++ app/src/main/res/layout/sign_in.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:id="@+id/first_title_main"
-    tools:context="com.cookandroid.teamproject1.id.view.SignInActivity"
+    tools:context="com.cookandroid.teamproject1.id.view.activity.SignInActivity"
     >
