+++ app/src/main/res/layout/sign_uping.xml
@@ -7,7 +7,7 @@
     android:background="@color/tlover_background_color"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context="com.cookandroid.teamproject1.id.view.SignUpingActivity">
+    tools:context="com.cookandroid.teamproject1.id.view.activity.SignUpingActivity">
 
 
     <TextView
